src: setting up dockerfile and docker build/push github actions pipeline
We create a dockerfile that runs the unit/integration tests and then builds the desired projects and starts them on X port. With this Dockerfile we can then create a CI/CD pipeline through Github Actions. Each commit/PR will build the Dockerfile and push the image to the container registry, where Azure will grab the image from and deploy when there's a change in the image version.